pyproject.toml
[project]
name = "django-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2",
    "django-configurations>=2.5.1",
    "django-guardian>=3.0.0",
    "django-ninja>=1.4.1",
    "django-redis>=5.4.0",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
]

[dependency-groups]
dev = [
    "ruff>=0.11.10",
]
